+++ rasa_nlu/featurizers/count_vectors_featurizer.py
@@ -25,7 +25,10 @@ if typing.TYPE_CHECKING:
 
 
 class CountVectorsFeaturizer(Featurizer):
-    """Bag of words featurizer"""
+    """Bag of words featurizer
+
+    Creates bag-of-words representation of intent features
+    using sklearn's `CountVectorizer`"""
 
     name = "intent_featurizer_count_vectors"
 
@@ -34,6 +37,9 @@ class CountVectorsFeaturizer(Featurizer):
     requires = []
 
     defaults = {
+        # regular expression for tokens
+        "token_pattern": r'(?u)\b\w\w+\b',
+
         # min number of word occurancies in the document to add to vocabulary
         "min_df": 1,
 
@@ -47,12 +53,13 @@ class CountVectorsFeaturizer(Featurizer):
     }
 
     def __init__(self, component_config=None):
-        # type: (RasaNLUModelConfig) -> None
         """Construct a new count vectorizer using the sklearn framework."""
 
         super(CountVectorsFeaturizer, self).__init__(component_config)
 
-        # default parameters
+        # regular expression for tokens
+        self.token_pattern = self.component_config['token_pattern']
+
         # min number of word occurancies in the document to add to vocabulary
         self.min_df = self.component_config['min_df']
 
@@ -79,8 +86,9 @@ class CountVectorsFeaturizer(Featurizer):
         from sklearn.feature_extraction.text import CountVectorizer
 
         # use even single character word as a token
-        self.vect = CountVectorizer(token_pattern=r"(?u)\b\w+\b",
-                                    ngram_range=(self.min_ngram, self.max_ngram),
+        self.vect = CountVectorizer(token_pattern=self.token_pattern,
+                                    ngram_range=(self.min_ngram,
+                                                 self.max_ngram),
                                     max_df=self.max_df,
                                     min_df=self.min_df)
 
@@ -99,7 +107,11 @@ class CountVectorsFeaturizer(Featurizer):
 
     def process(self, message, **kwargs):
         # type: (Message, **Any) -> None
-        if self.vect is not None:
+        if self.vect is None:
+            logger.error("There is no trained CountVectorizer: "
+                         "component is either not trained or "
+                         "didn't receive enough training data")
+        else:
             bag = self.vect.transform([self._lemmatize(message)]).toarray()
             message.set("text_features", bag)
 
@@ -111,11 +123,13 @@ class CountVectorsFeaturizer(Featurizer):
             return message.text
 
     @classmethod
-    def load(cls, model_dir=None,
-             model_metadata=None,
-             cached_component=None,
-             **kwargs):
-        # type: (Text, Metadata, Optional[Component], **Any) -> CountVectorsFeaturizer
+    def load(cls,
+             model_dir=None,  # type: Text
+             model_metadata=None,  # type: Metadata
+             cached_component=None,  # type: Optional[Component]
+             **kwargs  # type: **Any
+             ):
+        # type: (...) -> CountVectorsFeaturizer
 
         meta = model_metadata.for_component(cls.name)
 
@@ -124,8 +138,9 @@ class CountVectorsFeaturizer(Featurizer):
             featurizer_file = os.path.join(model_dir, file_name)
             return utils.pycloud_unpickle(featurizer_file)
         else:
-            raise Exception("Failed to load featurizer. Path {} "
-                            "doesn't exist".format(os.path.abspath(model_dir)))
+            logger.warning("Failed to load featurizer. Maybe path {} "
+                           "doesn't exist".format(os.path.abspath(model_dir)))
+            return CountVectorsFeaturizer(meta)
 
     def persist(self, model_dir):
         # type: (Text) -> Dict[Text, Any]